With the screenshot save system
This was an idea provided by Lox, a few months ago.

The idea is simple :
* Save the data in a texture
* Make it fullscreen
* Take a photo/screenshot

Then read the resulting image in another software.

Boy... You'd think that the biggest time sink would
be step 1... But it was STEP 2 !
"Make it fullscreen" was such a pain. At first, I went
with an orthographic camera and draw on top of the
player camera (Not the photo camera, the actual user
view).
It worked... until the user started to take a screenshot.
I don't understand how you can draw on top of the camera
the user is holding, without shaders.

Still, I gave the Particle Emitter a try, but no cigar.
It's pretty difficult to make a fullscreen billboard
particle, and the internet is not helping that much.

So, shader it is ! Quite frankly, we're talking about
beginner level of shader.
Make a ([-1,-1],[1,1]) quad, display it on front of
everything.
This is shader basics 101.... That can still fail because
Unity default quad have weird coordinates by default.
UGH.

So I made a quad generator, with user-input coordinates.
Then it worked ! (Still had to do a Y-flip for some reason).

I read back the data in another Unity project and it worked,
so the next step is to write a sample POC "importer" and
showcase it !
I'll also try to write a python script for Blender too, just
to make people realize "It's not only about Unity".

Signed-off-by: Voyage <voyage@miouyouyou.fr>