Wrapping up for a first release
First alpha, as always... There's still a lot to deal
with.

Still, it's working ! I was able to load the picture
from another project and recreate the scene created
in the world, so I'm ready to call this alpha 1.
Just need to setup some nice textures and redo the
UV of the wall elements, so that they fit nicely,
and we should be good to go.

Signed-off-by: Voyage <voyage@miouyouyou.fr>