Proper normals texture arrays !
At last !

With the help of RenderDoc, I was able to get to the
bottom of the situation. Turned out to be not calling
Texture2DArray with the right parameters (actually, not
using the right signature).

So now the Texture Array Generator script has been fixed,
made more simple also, and the arrays have been fixed.

Also I made sure that the pen is not rotated automatically
for VR users, since it's just annoying, while keeping
the automatic orientation for Desktop users, since not doing
it is actually annoying in Desktop.

I'll have to fix the other side, though, now (the Unity
plugin).

Signed-off-by: Voyage <voyage@miouyouyou.fr>